+++ module/entities/product_models.go
@@ -10,7 +10,7 @@ type Product struct {
 	ImageURL    string    `json:"imageUrl" db:"image_url"`
 	Stock       int       `json:"stock" db:"stock"`
 	Notes       string    `json:"notes" db:"notes"`
-	Price       float64   `json:"price" db:"price"`
+	Price       int       `json:"price" db:"price"`
 	Location    string    `json:"location" db:"location"`
 	IsAvailable bool      `json:"isAvailable" db:"is_available"`
 	CreatedAt   time.Time `json:"createdAt" db:"created_at"`
